Begun formatting Testing Spreadsheet
To begin evaluating the effectiveness of our AI tool, we will be calculating F1 scores using pseudo-TP, FP, and FN measurements from the comparison of the AI responses against ground truths. These scores and gathered data will all be stored in the new ChatData.xlsx spreadsheet and used for further analysis later on.